requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.2.0
blessed==1.20.0
boto3==1.34.140
botocore==1.34.140
CacheControl==0.14.0
cachetools==5.3.3
cement==2.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cohere==5.5.8
colorcet==3.1.0
colorlog==6.8.2
colorspacious==1.1.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dask==2024.5.2
dataclasses-json==0.6.7
datamapplot==0.3.0
datashader==0.16.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.110.3
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.15.4
firebase-admin==6.2.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
git-filter-repo==2.38.0
gitdb==4.0.11
GitPython==3.1.32
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.23.3
google-auth-httplib2==0.2.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
groq==0.8.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
importlib-metadata==7.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-google-community==1.0.5
langchain-groq==0.1.5
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.1
langgraph==0.0.65
langsmith==0.1.75
lazy-loader==0.4
llvmlite==0.42.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.33.0
openapi-schema-pydantic==1.2.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
param==2.1.0
parameterized==0.9.0
partd==1.4.2
pathspec==0.10.1
pdf2image==1.16.3
pillow==10.3.0
pinecone==4.0.0
pinecone-client==3.2.2
pip==23.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyct==0.5.0
pydantic==2.7.4
pydantic-core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pylabeladjust==0.1.13
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pynndescent==0.5.12
pyparsing==3.1.1
PyPDF2==3.0.1
Pyqtree==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-mock==1.12.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
semantic-router==0.0.48
semantic-version==2.8.5
setuptools==65.5.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240622
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
uritemplate==4.1.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.1.9
websockets==12.0
xarray==2024.5.0
yarl==1.9.4
zipp==3.19.2
phospho==0.3.30
